acceptance/eventlet: make sure apache is stopped
When testing upgrades, we have some cases where we apply the wsgi
manifest first and then eventlet. The problem is apache which is
started so ceilo-api cannot be started because ports are already in use.

So this patch installs Apache on eventlet manifest and makes sure
the service is stopped before trying to start ceilo-api with eventlet.

Change-Id: I4aae37fcbf60e61c0eb8385dedc950674661b592