Add required db subclass to README.md
This commit adds the ::ceilometer::db subclass to the example in README.md.
The README.md should contain a working pseudo-minimal example for basic
ceilometer installation. At some point the 'db' subclass was required
for basic usage.

Without this class applying the example in README.md is unusable,
erroring out with this:

> Error: Could not find dependency Class[Ceilometer::Db] for
> Service[ceilometer-api] at
> /etc/puppet/modules/ceilometer/manifests/api.pp:112

Change-Id: Ifd0577e23661bc00b98621a622c44c0a9af70a5d